# backend/docs/swagger.yaml
definitions:
...
  domain.Draft:
    properties:
      author:
        type: string
      category:
        type: string
      content:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  domain.DraftPageResult:
    properties:
      articles:
        items:
          $ref: '#/definitions/domain.Draft'
        type: array
      total:
        type: integer
    type: object
...
  domain.IsrSetting:
    properties:
      mode:
        type: string
    type: object
...
  domain.LinkItem:
    properties:
      desc:
        type: string
      logo:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
  domain.Static:
    properties:
      fileType:
        type: string
      meta: {}
      name:
        type: string
      realPath:
        type: string
      sign:
        type: string
      staticType:
        $ref: '#/definitions/domain.StaticType'
      storageType:
        $ref: '#/definitions/domain.StorageType'
      updatedAt:
        type: string
    type: object
  domain.StaticPageResult:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Static'
        type: array
      total:
        type: integer
    type: object
...
  domain.StaticType:
    enum:
    - img
    - customPage
    - ""
    type: string
    x-enum-varnames:
    - ImgStaticType
    - CustomPageStaticType
    - FileStaticType
  domain.StorageType:
    enum:
    - minio
    - local
    type: string
    x-enum-varnames:
    - MinioStorageType
    - LocalStorageType
...
info:
  contact: {}
paths:
...
  /api/admin/draft:
    get:
      consumes:
      - application/json
      description: get drafts by options
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: toListView
        in: query
        name: toListView
        type: boolean
      - description: category
        in: query
        name: category
        type: string
      - description: page
        in: query
        name: tags
        type: string
      - description: title
        in: query
        name: title
        type: integer
      - description: sortCreateAt
        in: query
        name: sortCreateAt
        type: integer
      - description: startTime
        in: query
        name: startTime
        type: integer
      - description: endTime
        in: query
        name: endTime
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DraftPageResult'
      summary: get drafts by options
      tags:
      - Draft
    post:
      consumes:
      - application/json
      description: crate draft
      parameters:
      - description: draft
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/domain.Draft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Draft'
      summary: crate draft
      tags:
      - Draft
  /api/admin/draft/:id:
    delete:
      consumes:
      - application/json
      description: crate draft
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: delete draft by id
      tags:
      - Draft
    get:
      consumes:
      - application/json
      description: get draft by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Draft'
      summary: get draft by id
      tags:
      - Draft
    put:
      consumes:
      - application/json
      description: update draft by id
      parameters:
      - description: draft
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/domain.Draft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update draft by id
      tags:
      - Draft
  /api/admin/draft/publish/:id:
    post:
      consumes:
      - application/json
      description: publish and update draft
      parameters:
      - description: draft
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/domain.Draft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Draft'
      summary: publish and update draft
      tags:
      - Draft
  /api/admin/img:
    get:
      consumes:
      - application/json
      description: get images by option
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: staticType
        in: query
        name: staticType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.StaticPageResult'
      summary: get images by option
      tags:
      - Static
...
  /api/admin/img/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload image
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: waterMarkText
        in: formData
        name: waterMarkText
        required: true
        type: string
      - description: withWaterMark
        in: formData
        name: withWaterMark
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: Upload image
      tags:
      - Static
  /api/admin/isr:
    get:
      consumes:
      - application/json
      description: get isr setting
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.IsrSetting'
      summary: get isr setting
      tags:
      - Isr
    post:
      consumes:
      - application/json
      description: active isr
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: active isr
      tags:
      - Isr
    put:
      consumes:
      - application/json
      description: update isr setting
      parameters:
      - description: isr
        in: body
        name: isr
        required: true
        schema:
          $ref: '#/definitions/domain.IsrSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Draft'
      summary: update isr setting
      tags:
      - Isr
...
  /api/admin/meta/link:
    get:
      consumes:
      - application/json
      description: get link by meta
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LinkItem'
            type: array
      summary: get link by meta
      tags:
      - Link
    post:
      consumes:
      - application/json
      description: create link
      parameters:
      - description: link
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/domain.LinkItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: create link
      tags:
      - Link
    put:
      consumes:
      - application/json
      description: update link
      parameters:
      - description: link
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/domain.LinkItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: update link
      tags:
      - Link
...
swagger: "2.0"